Use new transactions API and change all ints to int64
Most types of Roblox IDs are well past the int limit! Not sure why it was working at all...

+++ user/util.go
@@ -29,8 +29,8 @@ type getByUserIDAPIResponse struct {
 	Errors   []getByUserIDAPIErrors `json:"errors"`
 }
 
-func shortPoll(lastID int, interval time.Duration, fun func(int,
-	chan<- interface{}) (int, error), updates chan interface{},
+func shortPoll(lastID int64, interval time.Duration, fun func(int64,
+	chan<- interface{}) (int64, error), updates chan interface{},
 	errorChan chan<- error, cancel <-chan bool) {
 	var err error
 	ticker := time.NewTicker(interval)
